Using Var
using generic variables type and assigning the data type on the right is a standard coding practice.  In the past I have always defined the data type both on the left and the right.

In my opinion in the past this made the code more readable.  However in the long term this practice will clutter your code.