Default Constructors
Adding a default constructor.  Note here if we go back to the Program.cs the Cusotmer class has a red squiggly line.  Just like from MS Word when you have misspelled a word however, in this case it indicates that we are not using the correct syntax.

Note:  Here we do not have a constructor that accepts no arguments.  In order to fix we will need to assign the appropriate arguement(s)