Move Order.cs to new file
Here also we created a generic List of Orders in the Customer Class.  Think of this as a shopping cart that you are tracking Orders with for the customer.

Note:  The List<Order> will require you to add the namespace to your using statements.  You can use the quick actions shortcut to do this on the fly.