Inheritance
Now in the Program.cs we use the constructor that accepts an integer and a string.  Here we use inheritance inherit from the constructor that accepts only an integer which in turn inherits from the Customer class that has no arguments.

Note:  Thin of this like stairs.  Each step applies another set of instructions and you can go up one or down one from it.